% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Clustering with kohonen SOM or FastPG}
\description{
Functions for cell clustering, visualization, and projection using the
kohonen R package (self-organizing maps) or FastPG (fast graph-based
community detection).
}
\keyword{internal}
