% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{.get_Pipeline_class}
\alias{.get_Pipeline_class}
\title{Create an R7 Pipeline Class}
\usage{
.get_Pipeline_class()
}
\value{
The S7 class object for Pipeline.
}
\description{
Returns the S7 class object for \code{Pipeline}. Created once and cached.
}
\keyword{internal}
