% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sw_channel_get_fluor}
\alias{sw_channel_get_fluor}
\title{Get Fluorescent Channel Names}
\usage{
sw_channel_get_fluor(ff_or_fs)
}
\arguments{
\item{ff_or_fs}{A \code{flowFrame} or \code{flowSet} object.}
}
\value{
Character vector of fluorescent channel names.
}
\description{
Returns the names of fluorescent/marker channels from a \code{flowFrame} or
\code{flowSet}, excluding scatter (FSC/SSC) and Time channels.
}
