% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_pipeline_length}
\alias{sw_pipeline_length}
\title{Get Number of Steps in a Pipeline}
\usage{
sw_pipeline_length(pipeline)
}
\arguments{
\item{pipeline}{A \code{Pipeline} object.}
}
\value{
Integer.
}
\description{
Get Number of Steps in a Pipeline
}
