% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_pipeline_remove}
\alias{sw_pipeline_remove}
\title{Remove a Step from a Pipeline}
\usage{
sw_pipeline_remove(pipeline, name_or_index)
}
\arguments{
\item{pipeline}{A \code{Pipeline} object.}

\item{name_or_index}{Character name or integer index of the step to remove.}
}
\value{
A new \code{Pipeline} object with the step removed.
}
\description{
Returns a new pipeline with the specified step removed.
}
