% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_pipeline_step_filter_margins}
\alias{sw_pipeline_step_filter_margins}
\title{Create a Step for Margin Removal}
\usage{
sw_pipeline_step_filter_margins(
  ...,
  input_type = .SW_STEP_DEFAULT_TYPES$remove_margins$input,
  output_type = .SW_STEP_DEFAULT_TYPES$remove_margins$output
)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_filter_margins}.}

\item{input_type}{Character vector overriding the default input port type.}

\item{output_type}{Character vector overriding the default output port type.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_filter_margins}}. Defaults: input/output
\code{"flowFrame"}.
}
