% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_pipeline_step_names}
\alias{sw_pipeline_step_names}
\title{Get Step Names from a Pipeline}
\usage{
sw_pipeline_step_names(pipeline)
}
\arguments{
\item{pipeline}{A \code{Pipeline} object.}
}
\value{
Character vector of step names.
}
\description{
Get Step Names from a Pipeline
}
