% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_step_signal_qc}
\alias{sw_step_signal_qc}
\title{Create a Step for Signal QC}
\usage{
sw_step_signal_qc(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_signal_qc}.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_signal_qc}}.
}
