% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{sw_transform_apply}
\alias{sw_transform_apply}
\title{Apply Scale Transformations}
\usage{
sw_transform_apply(x, trans_list)
}
\arguments{
\item{x}{A \code{flowCore::flowFrame} or \code{flowCore::flowSet}.}

\item{trans_list}{A \code{flowCore::transformList} object, typically
obtained from \code{\link{sw_transform_estimate}}.}
}
\value{
The transformed \code{flowFrame} or \code{flowSet}.
}
\description{
Applies a pre-computed \code{flowCore::transformList} to a
\code{flowFrame} or \code{flowSet}.
}
\details{
Ported from CytoPipeline (UCLouvain-CBIO).
}
\examples{
\dontrun{
trans <- sw_transform_estimate(ff, fluo_method = "estimateLogicle")
ff_transformed <- sw_transform_apply(ff, trans)
}

}
