% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_step_batch_correct}
\alias{sw_step_batch_correct}
\title{Create a Step for Batch Correction}
\usage{
sw_step_batch_correct(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_batch_correct}.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_batch_correct}}.
}
