% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_step_cluster}
\alias{sw_step_cluster}
\title{Create a Step for Clustering}
\usage{
sw_step_cluster(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_cluster}.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_cluster}}.
}
