% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_step_normalize}
\alias{sw_step_normalize}
\title{Create a Step for Data Normalisation}
\usage{
sw_step_normalize(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_normalize}.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_normalize}}.
}
