% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composable_pipeline.R
\name{sw_step_remove_margins}
\alias{sw_step_remove_margins}
\title{Create a Step for Margin Removal}
\usage{
sw_step_remove_margins(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{sw_remove_margins}.}
}
\value{
A \code{ProcessingStep} object.
}
\description{
Convenience constructor for a \code{ProcessingStep} that wraps
\code{\link{sw_remove_margins}}.
}
